optimising RefineRegistry component
