+++ src/components/Navigation/HomeButton.vue
